switch to onMouseUp to avoid chrome SVG bug
On each virtual dom render, the devtools reveal that chrome is updating the internal structure of each `use` element. This seems to break click handlers and reset the click status.

This bug does not repro in Firefox. The Firefox devtools show the same updates, but the click handlers work as expected in Firefox.